# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alumni Influencers API",
        "description": "API documentation for the Alumni Influencers Bidding System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Alumni-Influencers",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/alumni/profile": {
            "get": {
                "tags": [
                    "Alumni Profile"
                ],
                "summary": "Get alumni profile",
                "operationId": "cc5fbcbfa5b162056df05b6aef73a6c9",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/alumni/profile/image": {
            "put": {
                "tags": [
                    "Alumni Profile"
                ],
                "summary": "Update profile image",
                "operationId": "328bdedb9fad771d211f8ffd9df7679d",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "profile_image_url": {
                                        "type": "string",
                                        "example": "http://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/alumni/profile/linkedin": {
            "put": {
                "tags": [
                    "Alumni Profile"
                ],
                "summary": "Update LinkedIn URL",
                "operationId": "f9fb8ad29ec5cd70fd34cc19a29506a1",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "linkedin_url": {
                                        "type": "string",
                                        "example": "https://linkedin.com/in/johndoe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/alumni/degrees": {
            "put": {
                "tags": [
                    "Alumni Profile"
                ],
                "summary": "Update degrees list (JSON)",
                "operationId": "09bb17f8696f160ac4262ed92b005acf",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "degrees": {
                                        "type": "string",
                                        "example": "['BSc CS', 'MSc AI']"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alumni Profile"
                ],
                "summary": "Add a new degree",
                "operationId": "bd1c8da99a6e355975dcdb63c6903ca0",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "degree": {
                                        "type": "string",
                                        "example": "BSc Computer Science"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Added"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alumni Profile"
                ],
                "summary": "Delete a degree by index",
                "operationId": "5e9c10a9648270b2445e7fb467961a0a",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "index": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Invalid index"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/alumni/profile/completion-status": {
            "get": {
                "tags": [
                    "Alumni Profile"
                ],
                "summary": "Get profile completion status",
                "operationId": "5d45b243aab82521d1cc0482890f6c11",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status returned"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "apikey",
                "in": "header"
            }
        }
    },
...
}